--- a/Steepfile
+++ b/Steepfile
@@ -1,0 +1,5 @@
+target :lib do
+  signature "sig"
+
+  check "lib"
+end