--- a/hashids.gemspec
+++ b/hashids.gemspec
@@ -1,7 +1,6 @@
 # encoding: utf-8
-lib = File.expand_path('../lib', __FILE__)
-$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'hashids'
+
+require_relative "lib/hashids/version"
 
 Gem::Specification.new do |gem|
   gem.name          = "hashids"