--- a/lib/hashids/version.rb
+++ b/lib/hashids/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Hashids
+  VERSION = "1.0.6"
+end