--- a/lib/hashids2.rb
+++ b/lib/hashids2.rb
@@ -151,15 +151,15 @@ class Hashids2
   end
 
   def consistent_shuffle(collection_to_shuffle, salt_part_1, salt_part_2, max_salt_length)
-    chars = collection_to_shuffle.dup
-
     salt_part_1_length = salt_part_1.length
 
-    return chars if collection_to_shuffle.empty? || max_salt_length == 0 || salt_part_1.nil? || salt_part_1_length == 0
+    return collection_to_shuffle if collection_to_shuffle.empty? || max_salt_length == 0 || salt_part_1.nil? || salt_part_1_length == 0
+
+    chars = collection_to_shuffle.dup
 
     idx = ord_total = 0
 
-    i = collection_to_shuffle.length-1
+    i = collection_to_shuffle.length - 1
     while i >= 1
       raise ArgumentError, "Salt is too short in shuffle" if idx >= salt_part_1_length && salt_part_2.nil?
       ord_total += n = (idx >= salt_part_1_length ? salt_part_2[idx - salt_part_1_length] : salt_part_1[idx]).ord