… lib/hashids3.rb
… spec/hashids2_spec.rb
… spec/hashids3_spec.rb