WIP, currently this is using a copy of `Hashids` to allow me to easily run both implementations side by side

Note: I will likely split this PR into smaller stacked ones shortly.

Some numbers (v1 is original implemation, v2 is the one Im working on):

```
Hashids#encode: 100 numbers
-------------------

Warming up --------------------------------------
                  v1   176.000  i/100ms
                  v2   297.000  i/100ms
Calculating -------------------------------------
                  v1      1.804k (± 1.9%) i/s -      5.456k in   3.025510s
                  v2      3.024k (± 1.8%) i/s -      9.207k in   3.045634s

Comparison:
                  v2:     3024.0 i/s
                  v1:     1804.0 i/s - 1.68x  slower



## Memory:

Calculating -------------------------------------
                  v1   410.384k memsize (     0.000  retained)
                         5.994k objects (     0.000  retained)
                        50.000  strings (     0.000  retained)
                  v2    11.024k memsize (     0.000  retained)
                       204.000  objects (     0.000  retained)
                        50.000  strings (     0.000  retained)

Comparison:
                  v2:      11024 allocated
                  v1:     410384 allocated - 37.23x more
```